Standardize the Setup Link + Directory sync
