+++ lib/ui/hooks/useDirectory.ts
@@ -3,8 +3,10 @@ import type { Directory } from '@boxyhq/saml-jackson';
 import type { ApiError, ApiSuccess } from 'types';
 import { fetcher } from '@lib/ui/utils';
 
-const useDirectory = (directoryId: string) => {
-  const url = `/api/admin/directory-sync/${directoryId}`;
+const useDirectory = (directoryId: string, setupLinkToken?: string) => {
+  const url = setupLinkToken
+    ? `/api/setup/${setupLinkToken}/directory-sync/${directoryId}`
+    : `/api/admin/directory-sync/${directoryId}`;
 
   const { data, error } = useSWR<ApiSuccess<Directory>, ApiError>(url, fetcher);
