+++ pages/admin/directory-sync/[directoryId]/events/index.tsx
@@ -71,7 +71,7 @@ const Events: NextPage = () => {
       <h2 className='mt-5 font-bold text-gray-700 md:text-xl'>{directory.name}</h2>
       <div className='w-full md:w-3/4'>
         <DirectoryTab directory={directory} activeTab='events' />
-        {events.length === 0 && paginate.offset ? (
+        {events.length === 0 && paginate.offset === 0 ? (
           <EmptyState title={t('no_webhook_events_found')} />
         ) : (
           <>
