+++ pages/api/setup/[token]/directory-sync/providers.ts
@@ -1,8 +1,7 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import jackson from '@lib/jackson';
-import { checkSession } from '@lib/middleware';
 
-export const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const { method } = req;
 
   switch (method) {
@@ -23,4 +22,4 @@ const handleGET = async (req: NextApiRequest, res: NextApiResponse) => {
   return res.json({ data: providers });
 };
 
-export default checkSession(handler);
+export default handler;
