+++ pages/setup/[token]/directory-sync/[directoryId]/edit.tsx
@@ -1,153 +1,203 @@
 import type { NextPage, GetServerSidePropsContext } from 'next';
-import React from 'react';
 import { useRouter } from 'next/router';
-import { useTranslation } from 'next-i18next';
-import jackson from '@lib/jackson';
-import { inferSSRProps } from '@lib/inferSSRProps';
-import { LinkBack } from '@components/LinkBack';
-import { ButtonPrimary } from '@components/ButtonPrimary';
-import { errorToast, successToast } from '@components/Toaster';
+import React from 'react';
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import EditDirectory from '@components/dsync/EditDirectory';
 
-const Edit: NextPage<inferSSRProps<typeof getServerSideProps>> = ({
-  directory: { id, name, log_webhook_events, webhook },
-}) => {
-  const { t } = useTranslation('common');
+const DirectoryEditPage: NextPage = () => {
   const router = useRouter();
-  const { token } = router.query;
-  const [directory, setDirectory] = React.useState({
-    name,
-    log_webhook_events,
-    webhook_url: webhook.endpoint,
-    webhook_secret: webhook.secret,
-  });
 
-  const [loading, setLoading] = React.useState(false);
+  const { token, directoryId } = router.query as { token: string; directoryId: string };
 
-  const onSubmit = async (event: React.FormEvent) => {
-    event.preventDefault();
-
-    setLoading(true);
-
-    const rawResponse = await fetch(`/api/setup/${token}/directory-sync/${id}`, {
-      method: 'PUT',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(directory),
-    });
-
-    setLoading(false);
-
-    const { data, error } = await rawResponse.json();
-
-    if (error) {
-      errorToast(error.message);
-      return;
-    }
-
-    if (data) {
-      successToast('Directory updated successfully');
-      router.replace(`/setup/${token}/directory-sync`);
-    }
-  };
-
-  const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const target = event.target as HTMLInputElement;
-    const value = target.type === 'checkbox' ? target.checked : target.value;
-
-    setDirectory({
-      ...directory,
-      [target.id]: value,
-    });
-  };
-
-  return (
-    <div>
-      <LinkBack href={`/setup/${token}/directory-sync`} />
-      <h2 className='mb-5 mt-5 font-bold text-gray-700 md:text-xl'>Update Directory</h2>
-      <div className='min-w-[28rem] rounded border border-gray-200 bg-white p-6 dark:border-gray-700 dark:bg-gray-800 md:w-3/4 md:max-w-lg'>
-        <form onSubmit={onSubmit}>
-          <div className='flex flex-col space-y-3'>
-            <div className='form-control w-full'>
-              <label className='label'>
-                <span className='label-text'>Directory name</span>
-              </label>
-              <input
-                type='text'
-                id='name'
-                className='input-bordered input w-full'
-                required
-                onChange={onChange}
-                value={directory.name}
-              />
-            </div>
-            <div className='form-control w-full'>
-              <label className='label'>
-                <span className='label-text'>Webhook URL</span>
-              </label>
-              <input
-                type='text'
-                id='webhook_url'
-                className='input-bordered input w-full'
-                onChange={onChange}
-                value={directory.webhook_url}
-              />
-            </div>
-            <div className='form-control w-full'>
-              <label className='label'>
-                <span className='label-text'>Webhook secret</span>
-              </label>
-              <input
-                type='text'
-                id='webhook_secret'
-                className='input-bordered input w-full'
-                onChange={onChange}
-                value={directory.webhook_secret}
-              />
-            </div>
-            <div className='form-control w-full py-2'>
-              <div className='flex items-center'>
-                <input
-                  id='log_webhook_events'
-                  type='checkbox'
-                  checked={directory.log_webhook_events}
-                  onChange={onChange}
-                  className='h-4 w-4 rounded border-gray-300 bg-gray-100 text-blue-600 focus:ring-2 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:ring-offset-gray-800 dark:focus:ring-blue-600'
-                />
-                <label className='ml-2 text-sm font-medium text-gray-900 dark:text-gray-300'>
-                  Enable Webhook events logging
-                </label>
-              </div>
-            </div>
-            <div>
-              <ButtonPrimary type='submit' loading={loading}>
-                {t('save_changes')}
-              </ButtonPrimary>
-            </div>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
+  return <EditDirectory directoryId={directoryId} setupLinkToken={token} />;
 };
 
 export const getServerSideProps = async (context: GetServerSidePropsContext) => {
-  const { directoryId } = context.query;
-  const { directorySyncController } = await jackson();
-
-  const { data: directory } = await directorySyncController.directories.get(directoryId as string);
-
-  if (!directory) {
-    return {
-      notFound: true,
-    };
-  }
+  const { locale } = context;
 
   return {
     props: {
-      directory,
+      ...(locale ? await serverSideTranslations(locale, ['common']) : {}),
     },
   };
 };
 
-export default Edit;
+export default DirectoryEditPage;
+
+// import type { NextPage, GetServerSidePropsContext } from 'next';
+// import React from 'react';
+// import { useRouter } from 'next/router';
+// import { useTranslation } from 'next-i18next';
+// import { LinkBack } from '@components/LinkBack';
+// import { ButtonPrimary } from '@components/ButtonPrimary';
+// import { errorToast, successToast } from '@components/Toaster';
+// import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+
+// import type { ApiResponse } from 'types';
+// import Loading from '@components/Loading';
+// import useDirectory from '@lib/ui/hooks/useDirectory';
+
+// type FormState = Pick<Directory, 'name' | 'log_webhook_events'> & {
+//   webhook_url: string;
+//   webhook_secret: string;
+// };
+
+// const defaultFormState: FormState = {
+//   name: '',
+//   webhook_url: '',
+//   webhook_secret: '',
+//   log_webhook_events: false,
+// };
+
+// const DirectoryEditPage: NextPage = () => {
+//   const { t } = useTranslation('common');
+//   const router = useRouter();
+
+//   const [directory, setDirectory] = React.useState<FormState>(defaultFormState);
+//   const [loading, setLoading] = React.useState(false);
+
+//   const { token, directoryId } = router.query as { token: string; directoryId: string };
+
+//   const { directory: data, isLoading, error } = useDirectory(directoryId);
+
+//   React.useEffect(() => {
+//     if (data) {
+//       const directory = data;
+
+//       setDirectory({
+//         name: directory.name,
+//         webhook_url: directory.webhook.endpoint,
+//         webhook_secret: directory.webhook.secret,
+//         log_webhook_events: directory.log_webhook_events,
+//       });
+//     }
+//   }, [data]);
+
+//   if (isLoading) {
+//     return <Loading />;
+//   }
+
+//   if (error) {
+//     errorToast(error.message);
+//     return null;
+//   }
+
+//   const onSubmit = async (event: React.FormEvent) => {
+//     event.preventDefault();
+
+//     setLoading(true);
+
+//     const rawResponse = await fetch(`/api/setup/${token}/directory-sync/${directoryId}`, {
+//       method: 'PUT',
+//       headers: {
+//         'Content-Type': 'application/json',
+//       },
+//       body: JSON.stringify(directory),
+//     });
+
+//     setLoading(false);
+
+//     const { data, error } = await rawResponse.json();
+
+//     if (error) {
+//       errorToast(error.message);
+//       return;
+//     }
+
+//     if (data) {
+//       successToast(t('directory_updated_successfully'));
+//       router.replace(`/setup/${token}/directory-sync`);
+//     }
+//   };
+
+//   const onChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+//     const target = event.target as HTMLInputElement;
+//     const value = target.type === 'checkbox' ? target.checked : target.value;
+
+//     setDirectory({
+//       ...directory,
+//       [target.id]: value,
+//     });
+//   };
+
+//   return (
+//     <div>
+//       <LinkBack href={`/setup/${token}/directory-sync`} />
+//       <h2 className='mb-5 mt-5 font-bold text-gray-700 md:text-xl'>Update Directory</h2>
+//       <div className='min-w-[28rem] rounded border border-gray-200 bg-white p-6 dark:border-gray-700 dark:bg-gray-800 md:w-3/4 md:max-w-lg'>
+//         <form onSubmit={onSubmit}>
+//           <div className='flex flex-col space-y-3'>
+//             <div className='form-control w-full'>
+//               <label className='label'>
+//                 <span className='label-text'>Directory name</span>
+//               </label>
+//               <input
+//                 type='text'
+//                 id='name'
+//                 className='input-bordered input w-full'
+//                 required
+//                 onChange={onChange}
+//                 value={directory.name}
+//               />
+//             </div>
+//             <div className='form-control w-full'>
+//               <label className='label'>
+//                 <span className='label-text'>Webhook URL</span>
+//               </label>
+//               <input
+//                 type='text'
+//                 id='webhook_url'
+//                 className='input-bordered input w-full'
+//                 onChange={onChange}
+//                 value={directory.webhook_url}
+//               />
+//             </div>
+//             <div className='form-control w-full'>
+//               <label className='label'>
+//                 <span className='label-text'>Webhook secret</span>
+//               </label>
+//               <input
+//                 type='text'
+//                 id='webhook_secret'
+//                 className='input-bordered input w-full'
+//                 onChange={onChange}
+//                 value={directory.webhook_secret}
+//               />
+//             </div>
+//             <div className='form-control w-full py-2'>
+//               <div className='flex items-center'>
+//                 <input
+//                   id='log_webhook_events'
+//                   type='checkbox'
+//                   checked={directory.log_webhook_events}
+//                   onChange={onChange}
+//                   className='h-4 w-4 rounded border-gray-300 bg-gray-100 text-blue-600 focus:ring-2 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:ring-offset-gray-800 dark:focus:ring-blue-600'
+//                 />
+//                 <label className='ml-2 text-sm font-medium text-gray-900 dark:text-gray-300'>
+//                   {t('enable_webhook_events_logging')}
+//                 </label>
+//               </div>
+//             </div>
+//             <div>
+//               <ButtonPrimary type='submit' loading={loading}>
+//                 {t('save_changes')}
+//               </ButtonPrimary>
+//             </div>
+//           </div>
+//         </form>
+//       </div>
+//     </div>
+//   );
+// };
+
+// export const getServerSideProps = async (context: GetServerSidePropsContext) => {
+//   const { locale } = context;
+
+//   return {
+//     props: {
+//       ...(locale ? await serverSideTranslations(locale, ['common']) : {}),
+//     },
+//   };
+// };
+
+// export default DirectoryEditPage;
